tBasic upgraded

+++ src/env_68k/include/vga_video.h
@@ -53,7 +53,7 @@ ALL OF THIS REGISTER IS NOT USED NOW
 */
 
 void run_cmd(uint8_t cmd);
-void gotoxy(uint8_t col,uint8_t row);
+void gotoxy(uint16_t col,uint16_t row);
 void gohome();
 void setcolor(uint8_t txtcolor,uint8_t bgcolor);
 void clrscr();
